docs: a better name for float type

+++ src/main/java/argparser/ValueType.java
@@ -6,7 +6,7 @@ public enum ValueType {
         "^-?\\d+$"
     ),
     FLOAT(
-        "Decimal",
+        "Float",
         "^-?\\d+(\\.\\d*)?$"
     ),
     STRING(
